==> tools/50-load-compare-bench.R <==

source('tools/30-tpch-export.R')
source('tools/31-tpch-load.R')
source('tools/32-tpch-run.R')
source('tools/33-tpch-compare.R')
source('tools/34-tpch-bench.R')
source('tools/35-tpch-bench-dplyr.R')
source('tools/36-tpch-bench-compare.R')
