==> tools/99-sync.R <==
source("tools/01-dplyr-methods.R", echo = TRUE)
source("tools/02-duckplyr_df-methods.R", echo = TRUE)
source("tools/03-tests.R", echo = TRUE)
source("tools/04-dplyr-tests.R", echo = TRUE)
source("tools/05-duckdb-tests.R", echo = TRUE)
source("tools/06-patch-duckdb.R", echo = TRUE)
source("tools/30-tpch-export.R", echo = TRUE)
source("tools/31-tpch-load-qs.R", echo = TRUE)
source("tools/37-tpch-peel.R", echo = TRUE)
source("tools/39-tpch-peel-oo.R", echo = TRUE)
